pyproject.toml
[tool.poetry]
name = "nli-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "2.29.0"
numpy = "^1.24.3"
logzero = "^1.7.0"
toolz = "^0.12.0"
pyyaml = "^6.0"
prettytable = "^3.7.0"
seqeval = "^1.2.2"
mojimoji = "^0.0.12"
pytorch-warmup = "^0.1.1"
transformers = "4.8.0"
datasets = "1.8.0"
accelerate = "0.3.0"
wandb = "0.15.5"

[tool.poetry.dev-dependencies]
types-PyYAML = "^6.0.5"
isort = "^5.10.1"
black = "^22.3.0"
ipdb = "^0.13.9"
pytest = "^7.1.1"
mypy = "^0.942"
types-termcolor = "^1.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
